+++ src/include/sll/_sll_internal.h
@@ -9,7 +9,6 @@
 #include <sll/platform.h>
 #include <sll/string.h>
 #include <sll/types.h>
-#include <signal.h>
 #include <stdint.h>