+++ src/include/sll/api/string.h
@@ -6,14 +6,6 @@
 
 
 
-__SLL_EXTERNAL void sll_object_to_string(const sll_runtime_object_t*const* a,sll_array_length_t al,sll_string_t* o);
-
-
-
-__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_string_length_t sll_object_to_string_length(const sll_runtime_object_t*const* a,sll_array_length_t al,sll_bool_t q);
-
-
-
 /*~~DOCS~~
 TYPE api compilation_call optimizable var_arg
 FUNC sll_api_string_convert
@@ -22,17 +14,6 @@ ARG O Docs!
 RET Z Failure
 RET S Success*/
 __API_FUNC_DECL(string_convert);
-
-
-
-/*~~DOCS~~
-TYPE api compilation_call optimizable var_arg
-FUNC sll_api_string_length
-DESC Docs!
-ARG O Docs!
-RET 0 Failure
-RET I Success*/
-__API_FUNC_DECL(string_length);