+++ src/sll/api/string_api.c
@@ -5,6 +5,7 @@
 #include <sll/common.h>
 #include <sll/gc.h>
 #include <sll/handle.h>
+#include <sll/memory.h>
 #include <sll/runtime_object.h>
 #include <sll/static_object.h>
 #include <sll/string.h>
@@ -15,284 +16,221 @@
 
 
 
-#define INT_SIZE(v,o) \
-	do{ \
-		(o)++; \
-		(v)/=10; \
-	} while (v);
-
-
-
-static sll_string_length_t _write_int(uint64_t v,sll_string_length_t i,sll_string_t* o){
-	uint8_t k=0;
+static sll_string_length_t _write_int(uint64_t v,sll_string_t* o){
+	uint8_t i=0;
 	char bf[20];
 	do{
-		bf[k]=v%10;
-		k++;
+		bf[i]=v%10;
+		i++;
 		v/=10;
 	} while (v);
-	while (k){
-		k--;
-		o->v[i]=bf[k]+48;
-		i++;
+	sll_string_increase(o,i);
+	while (i){
+		i--;
+		o->v[o->l]=bf[i]+48;
+		o->l++;
 	}
 	return i;
 }
 
 
-static sll_string_length_t _write_char(sll_char_t c,sll_string_length_t i,sll_string_t* o){
+static void _write_char(sll_char_t c,sll_string_t* o){
 	if (c=='\"'||c=='\''||c=='\\'){
-		o->v[i]='\\';
-		o->v[i+1]=c;
-		return i+2;
+		sll_string_increase(o,2);
+		o->v[o->l]='\\';
+		o->v[o->l+1]=c;
+		o->l+=2;
 	}
 	else if (c=='\t'){
-		o->v[i]='\\';
-		o->v[i+1]='t';
-		return i+2;
+		sll_string_increase(o,2);
+		o->v[o->l]='\\';
+		o->v[o->l+1]='t';
+		o->l+=2;
 	}
 	else if (c=='\n'){
-		o->v[i]='\\';
-		o->v[i+1]='n';
-		return i+2;
+		sll_string_increase(o,2);
+		o->v[o->l]='\\';
+		o->v[o->l+1]='n';
+		o->l+=2;
 	}
 	else if (c=='\v'){
-		o->v[i]='\\';
-		o->v[i+1]='v';
-		return i+2;
+		sll_string_increase(o,2);
+		o->v[o->l]='\\';
+		o->v[o->l+1]='v';
+		o->l+=2;
 	}
 	else if (c=='\f'){
-		o->v[i]='\\';
-		o->v[i+1]='f';
-		return i+2;
+		sll_string_increase(o,2);
+		o->v[o->l]='\\';
+		o->v[o->l+1]='f';
+		o->l+=2;
 	}
 	else if (c=='\r'){
-		o->v[i]='\\';
-		o->v[i+1]='r';
-		return i+2;
+		sll_string_increase(o,2);
+		o->v[o->l]='\\';
+		o->v[o->l+1]='r';
+		o->l+=2;
 	}
 	else if (SLL_STRING_HEX_ESCAPE(c)){
-		o->v[i]='\\';
-		o->v[i+1]='x';
-		o->v[i+2]=(c>>4)+((c>>4)>9?87:48);
-		o->v[i+3]=(c&15)+((c&15)>9?87:48);
-		return i+4;
+		sll_string_increase(o,4);
+		o->v[o->l]='\\';
+		o->v[o->l+1]='x';
+		o->v[o->l+2]=(c>>4)+((c>>4)>9?87:48);
+		o->v[o->l+3]=(c&15)+((c&15)>9?87:48);
+		o->l+=4;
 	}
-	o->v[i]=c;
-	return i+1;
+	else{
+		sll_string_increase(o,1);
+		o->v[o->l]=c;
+		o->l++;
+	}
 }
 
 
 
-static sll_string_length_t _object_to_string(const sll_runtime_object_t* a,sll_bool_t q,sll_string_t* o,sll_string_index_t i){
+static void _object_to_string(const sll_runtime_object_t* a,sll_bool_t q,sll_string_t* o){
 	if (!a->rc){
-		return SLL_STRING_INSERT_POINTER_STATIC("<released object>",i,o);
+		sll_string_increase(o,19);
+		sll_copy_string(SLL_CHAR("<released object>"),o->v+o->l);
+		o->l+=19;
+		return;
 	}
 	switch (SLL_RUNTIME_OBJECT_GET_TYPE(a)){
 		case SLL_RUNTIME_OBJECT_TYPE_INT:
 			{
 				sll_integer_t v=a->dt.i;
 				if (v<0){
-					o->v[i]='-';
+					sll_string_increase(o,1);
+					o->v[o->l]='-';
 					v=-v;
-					i++;
+					o->l++;
 				}
-				return _write_int(v,i,o);
+				_write_int(v,o);
+				return;
 			}
 		case SLL_RUNTIME_OBJECT_TYPE_FLOAT:
-			return i+snprintf((char*)(o->v+i),4096,"%.16lf",a->dt.f);
+			{
+				sll_char_t bf[256];
+				sll_string_length_t bfl=snprintf((char*)bf,256,"%.16lf",a->dt.f);
+				sll_string_increase(o,bfl);
+				sll_copy_data(bf,bfl,o->v+o->l);
+				o->l+=bfl;
+				return;
+			}
 		case SLL_RUNTIME_OBJECT_TYPE_CHAR:
 			if (q){
-				o->v[i]='\'';
-				i=_write_char(a->dt.c,i+1,o);
-				o->v[i]='\'';
-				return i+1;
+				sll_string_increase(o,1);
+				o->v[o->l]='\'';
+				o->l++;
+				_write_char(a->dt.c,o);
+				sll_string_increase(o,1);
+				o->v[o->l]='\'';
+				o->l++;
+				return;
 			}
-			o->v[i]=a->dt.c;
-			return i+1;
+			sll_string_increase(o,1);
+			o->v[o->l]=a->dt.c;
+			o->l++;
+			return;
 		case SLL_RUNTIME_OBJECT_TYPE_STRING:
 			if (q){
-				o->v[i]='\"';
-				i++;
-				for (sll_string_length_t j=0;j<a->dt.s.l;j++){
-					i=_write_char(a->dt.s.v[j],i,o);
+				sll_string_increase(o,1);
+				o->v[o->l]='\"';
+				o->l++;
+				for (sll_string_length_t i=0;i<a->dt.s.l;i++){
+					_write_char(a->dt.s.v[i],o);
 				}
-				o->v[i]='\"';
-				i++;
-				return i;
+				sll_string_increase(o,1);
+				o->v[o->l]='\"';
+				o->l++;
+				return;
 			}
-			sll_copy_data(a->dt.s.v,a->dt.s.l,o->v+i);
-			return i+a->dt.s.l;
+			sll_string_increase(o,a->dt.s.l);
+			sll_copy_data(a->dt.s.v,a->dt.s.l,o->v+o->l);
+			o->l+=a->dt.s.l;
+			return;
 		case SLL_RUNTIME_OBJECT_TYPE_ARRAY:
-			o->v[i]='[';
-			i++;
-			for (sll_array_length_t k=0;k<a->dt.a.l;k++){
-				if (k){
-					o->v[i]=' ';
-					i++;
+			sll_string_increase(o,1);
+			o->v[o->l]='[';
+			o->l++;
+			for (sll_array_length_t i=0;i<a->dt.a.l;i++){
+				if (i){
+					sll_string_increase(o,1);
+					o->v[o->l]=' ';
+					o->l++;
 				}
-				i=_object_to_string(*(a->dt.a.v+k),1,o,i);
+				_object_to_string(*(a->dt.a.v+i),1,o);
 			}
-			o->v[i]=']';
-			return i+1;
+			sll_string_increase(o,1);
+			o->v[o->l]=']';
+			o->l++;
+			return;
 		case SLL_RUNTIME_OBJECT_TYPE_HANDLE:
 			{
 				sll_handle_descriptor_t* hd=(sll_current_runtime_data?SLL_HANDLE_LOOKUP_DESCRIPTOR(sll_current_runtime_data->hl,a->dt.h.t):NULL);
-				sll_handle_t v=a->dt.h.h;
 				if (hd&&hd->sf){
-					return hd->sf(v,i,o);
+					hd->sf(a->dt.h.h,o);
+					return;
 				}
-				i=_write_int(a->dt.h.h,i,o);
-				o->v[i]='$';
-				i++;
+				_write_int(a->dt.h.h,o);
+				sll_string_increase(o,1);
+				o->v[o->l]='$';
+				o->l++;
 				if (hd){
-					sll_copy_data(hd->nm,hd->nml,o->v+i);
-					return i+hd->nml;
+					sll_string_increase(o,hd->nml);
+					sll_copy_data(hd->nm,hd->nml,o->v+o->l);
+					o->l+=hd->nml;
+					return;
 				}
-				return _write_int(a->dt.h.t,i,o);
+				_write_int(a->dt.h.t,o);
+				return;
 			}
 		case SLL_RUNTIME_OBJECT_TYPE_MAP:
-			o->v[i]='<';
-			i++;
-			for (sll_map_length_t k=0;k<(a->dt.m.l<<1);k++){
-				if (k){
-					o->v[i]=' ';
-					i++;
+			sll_string_increase(o,1);
+			o->v[o->l]='<';
+			o->l++;
+			for (sll_array_length_t i=0;i<(a->dt.m.l<<1);i++){
+				if (i){
+					sll_string_increase(o,1);
+					o->v[o->l]=' ';
+					o->l++;
 				}
-				i=_object_to_string(*(a->dt.m.v+k),1,o,i);
+				_object_to_string(*(a->dt.m.v+i),1,o);
 			}
-			o->v[i]='>';
-			return i+1;
+			sll_string_increase(o,1);
+			o->v[o->l]='>';
+			o->l++;
+			return;
 		case RUNTIME_OBJECT_TYPE_FUNCTION_ID:
-			o->v[i]='#';
-			return _write_int(a->dt.i,i+1,o);
+			sll_string_increase(o,1);
+			o->v[o->l]='#';
+			o->l++;
+			_write_int(a->dt.i,o);
+			return;
 		case RUNTIME_OBJECT_TYPE_UNKNOWN:
-			return SLL_STRING_INSERT_POINTER_STATIC("unknown",i,o);
 		default:
-			SLL_UNREACHABLE();
+			sll_string_increase(o,9);
+			sll_copy_string(SLL_CHAR("<unknown>"),o->v+o->l);
+			o->l+=9;
+			return;
 	}
-	return 0;
-}
-
-
-
-__SLL_EXTERNAL void sll_object_to_string(const sll_runtime_object_t*const* a,sll_array_length_t al,sll_string_t* o){
-	if (!al){
-		SLL_INIT_STRING(o);
-		return;
-	}
-	sll_string_create(sll_object_to_string_length(a,al,0),o);
-	sll_string_length_t i=0;
-	for (sll_array_length_t j=0;j<al;j++){
-		i=_object_to_string(*(a+j),0,o,i);
-	}
-	SLL_ASSERT(i==o->l);
-	sll_string_calculate_checksum(o);
-}
-
-
-
-__SLL_EXTERNAL sll_string_length_t sll_object_to_string_length(const sll_runtime_object_t*const* al,sll_array_length_t all,sll_bool_t q){
-	sll_string_length_t o=0;
-	for (sll_array_length_t i=0;i<all;i++){
-		const sll_runtime_object_t* a=*(al+i);
-		if (!a->rc){
-			o+=17;
-			continue;
-		}
-		switch (SLL_RUNTIME_OBJECT_GET_TYPE(a)){
-			case SLL_RUNTIME_OBJECT_TYPE_INT:
-				{
-					sll_integer_t v=a->dt.i;
-					if (v<0){
-						o++;
-						v=-v;
-					}
-					INT_SIZE(v,o);
-					break;
-				}
-			case SLL_RUNTIME_OBJECT_TYPE_FLOAT:
-				o+=snprintf(NULL,0,"%.16lf",a->dt.f);
-				break;
-			case SLL_RUNTIME_OBJECT_TYPE_CHAR:
-				o++;
-				if (q){
-					o+=2;
-					if (SLL_STRING_ESCAPE(a->dt.c)){
-						o++;
-					}
-					else if (SLL_STRING_HEX_ESCAPE(a->dt.c)){
-						o+=3;
-					}
-				}
-				break;
-			case SLL_RUNTIME_OBJECT_TYPE_STRING:
-				o+=a->dt.s.l;
-				if (q){
-					o+=2;
-					for (sll_string_length_t j=0;j<a->dt.s.l;j++){
-						if (SLL_STRING_ESCAPE(a->dt.s.v[j])){
-							o++;
-						}
-						else if (SLL_STRING_HEX_ESCAPE(a->dt.s.v[j])){
-							o+=3;
-						}
-					}
-				}
-				break;
-			case SLL_RUNTIME_OBJECT_TYPE_ARRAY:
-				o+=sll_object_to_string_length((const sll_runtime_object_t*const*)a->dt.a.v,a->dt.a.l,1)+a->dt.a.l+(!a->dt.a.l)+1;
-				break;
-			case SLL_RUNTIME_OBJECT_TYPE_HANDLE:
-				{
-					sll_handle_descriptor_t* hd=(sll_current_runtime_data?SLL_HANDLE_LOOKUP_DESCRIPTOR(sll_current_runtime_data->hl,a->dt.h.t):NULL);
-					sll_handle_t v=a->dt.h.h;
-					if (hd&&hd->sf){
-						o+=hd->sf(v,0,NULL);
-					}
-					else{
-						INT_SIZE(v,o);
-						o++;
-						if (hd){
-							o+=hd->nml;
-						}
-						else{
-							v=a->dt.h.t;
-							INT_SIZE(v,o);
-						}
-					}
-					break;
-				}
-			case SLL_RUNTIME_OBJECT_TYPE_MAP:
-				o+=sll_object_to_string_length((const sll_runtime_object_t*const*)a->dt.m.v,a->dt.m.l<<1,1)+(a->dt.m.l<<1)+(!a->dt.m.l)+1;
-				break;
-			case RUNTIME_OBJECT_TYPE_FUNCTION_ID:
-				{
-					sll_integer_t v=a->dt.i;
-					o++;
-					INT_SIZE(v,o);
-					break;
-				}
-			case RUNTIME_OBJECT_TYPE_UNKNOWN:
-				o+=9;
-				break;
-			default:
-				SLL_UNREACHABLE();
-		}
-	}
-	return o;
+	SLL_UNREACHABLE();
 }
 
 
 
 __API_FUNC(string_convert){
-	sll_object_to_string(a,ac,out);
-}
-
-
-
-__API_FUNC(string_length){
-	return sll_object_to_string_length(a,ac,0);
+	if (!ac){
+		SLL_INIT_STRING(out);
+		return;
+	}
+	sll_string_create(0,out);
+	out->v=sll_memory_move(out->v,SLL_MEMORY_MOVE_DIRECTION_TO_STACK);
+	for (sll_array_length_t i=0;i<ac;i++){
+		_object_to_string(*(a+i),0,out);
+	}
+	out->v=sll_memory_move(out->v,SLL_MEMORY_MOVE_DIRECTION_FROM_STACK);
+	sll_string_calculate_checksum(out);
 }